% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weighted.R
\name{fit_weighted}
\alias{fit_weighted}
\title{Standalone function to perform weighted fitting of a linear model}
\usage{
fit_weighted(
  weights,
  datasets,
  tol = 1e-06,
  family = c("linear", "logistic", "cox")
)
}
\arguments{
\item{weights}{Integration weights given to datasets}

\item{datasets}{A list of datasets (each dataset having design matrix "X", response "y",
potentially censoring vector "censor").
It is assumed the first indexed dataset is the local dataset.}

\item{tol}{Convergence criteria for logistic and Cox families: stops when the vector 2-norm of parameter estimates
differs by less than tol}

\item{family}{The model to be fit, one of "linear", "logistic", or "cox".}
}
\value{
A set of fitted post-integration linear model parameters
}
\description{
Standalone function to perform weighted fitting of a linear model
}
