% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weighted_linear.R
\name{fit_weighted_linear}
\alias{fit_weighted_linear}
\title{Function to perform weighted fitting of a linear model}
\usage{
fit_weighted_linear(weights, datasets, precomputed.data = NULL, trace = F)
}
\arguments{
\item{weights}{Integration weights given to datasets}

\item{datasets}{A list of datasets (each dataset having design matrix X, response y).
It is assumed the first indexed dataset is the local dataset.}

\item{precomputed.data}{Output from the precompute_linear function if already computed}
}
\value{
A set of fitted post-integration linear model parameters
}
\description{
Function to perform weighted fitting of a linear model
}
