% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weighted_survival.R
\name{fit_weighted_survival}
\alias{fit_weighted_survival}
\title{Function to perform weighted fitting of a Cox model}
\usage{
fit_weighted_survival(weights, datasets, trace = F)
}
\arguments{
\item{weights}{Integration weights given to datasets}

\item{datasets}{A list of datasets (each dataset having design matrix x, response y, and censoring vector censor).
It is assumed the first indexed dataset is the local dataset.}
}
\value{
A set of fitted post-integration Cox model parameters
}
\description{
Function to perform weighted fitting of a Cox model
}
