% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{internal_functions}
\alias{internal_functions}
\alias{precompute_linear}
\alias{fit_weighted_linear}
\alias{fit_weighted_logistic}
\alias{fit_weighted_survival}
\alias{test_coefs_linear}
\alias{test_coefs_logistic}
\alias{test_coefs_survival}
\alias{loocv_weighted_linear}
\alias{loocv_weighted_logistic}
\alias{loocv_weighted_survival}
\alias{loocv_weighted_2param}
\alias{loocv_weighted_logistic_neg}
\alias{loocv_weighted_logistic_neg_2param}
\alias{loocv_weighted_survival_neg}
\alias{loocv_weighted_survival_neg_2param}
\title{Internal fitting and leave-one-out functions for joint fitting and simulation functions}
\usage{
precompute_linear(datasets, trace = F)

fit_weighted_linear(weights, datasets, precomputed.data = NULL)

fit_weighted_logistic(weights, datasets, tol = 1e-06, trace = F)

fit_weighted_survival(weights, datasets, tol = 1e-06, trace = F)

test_coefs_linear(datasets, trace = F)

test_coefs_logistic(datasets, trace = F)

test_coefs_survival(datasets, trace = F)

loocv_weighted_linear(weights, datasets, precomputed.data = NULL)

loocv_weighted_logistic(weights, datasets, betafitted)

loocv_weighted_survival(weights, datasets, betafitted)

loocv_weighted_2param(alphabeta, datasets, precomputed.data = NULL)

loocv_weighted_logistic_neg(weights, datasets)

loocv_weighted_logistic_neg_2param(alphabeta, datasets)

loocv_weighted_survival_neg(weights, datasets)

loocv_weighted_survival_neg_2param(alphabeta, datasets)
}
\arguments{
\item{datasets}{A list of K datasets (each dataset having design matrix x, response y,
and potenailly censoring vector censor). It is assumed the first indexed dataset is the local dataset.}

\item{weights}{numeric vector of weights}

\item{precomputed.data}{Precomputed data for PRESS statistic calculation, from precompute_linear}

\item{tol}{Convergence criteria for logistic and Cox families: stops when the vector 2-norm of parameter estimates
differs by less than tol}

\item{betafitted}{Fitted vector of parameter estimates, likely from fit_weighted functions}

\item{alphabeta}{2-vector of p-value cutoffs for integration weights}
}
\description{
Internal fitting and leave-one-out functions for joint fitting and simulation functions
}
