% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_fitting.R
\name{joint_fitting}
\alias{joint_fitting}
\title{A joint fitting function to calculate integration weights for the full optimization and
2-parameter optimization methods as well as their post-integration weighted parameter
estimates.}
\usage{
joint_fitting(
  datasets,
  family = c("linear", "logistic", "cox"),
  method = c("opt", "testingopt"),
  trace = F
)
}
\arguments{
\item{datasets}{A list of K datasets (each dataset having design matrix x, response y, and censoring vector censor).
It is assumed the first indexed dataset is the local dataset.}

\item{family}{The model to be fit, one of "linear", "logistic", or "cox"}

\item{method}{One of:
\itemize{
\item opt - finds the length K-1 vector of integration weights maximizing the cross-validated partial likelihood
\item testingopt - using the better of testingpval or testingbinary as a starting point
find the 2-vector of p-value cutpoints that maximize the CVPL, turning a K-1 dimensional problem
into a 2 dimensional problem. For further details, refer to the loocv_weighted_survival_neg_2param
documentation
}}

\item{trace}{if TRUE, additionally prints family and method along with best weights}
}
\value{
A list of integration weights minimizing the CVPL,
the CVPL, the post-integration estimates, and the local-only estimates assuming no integration.
}
\description{
A joint fitting function to calculate integration weights for the full optimization and
2-parameter optimization methods as well as their post-integration weighted parameter
estimates.
}
