% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_fitting_linear.R
\name{joint_fitting_linear}
\alias{joint_fitting_linear}
\title{A joint fitting function to calculate integration weights for the full optimization and
2-parameter optimization methods as well as their post-integration weighted parameter
estimates.}
\usage{
joint_fitting_linear(
  datasets,
  fit.weighted = fit_weighted_linear,
  loocv.weighted = loocv_weighted_linear,
  precompute = precompute_linear,
  test.coefs = test_coefs_linear,
  method = c("opt", "testingopt"),
  trace = F
)
}
\arguments{
\item{datasets}{A list of K datasets (each dataset having design matrix x, response y, and censoring vector censor).
It is assumed the first indexed dataset is the local dataset.}

\item{fit.weighted}{The fit_weighted_linear function}

\item{loocv.weighted}{The loocv_weighted_linear function}

\item{test.coefs}{The test_coefs_linear function}

\item{method}{One of:
\itemize{
\item opt - finds the length K-1 vector of integration weights minimizing the PRESS statistic
\item testingopt - using the better of testingpval or testingbinary as a starting point
find the 2-vector of p-value cutpoints that maximize the CVPL, turning a K-1 dimensional problem
into a 2 dimensional problem. For further details, refer to the loocv_weighted_2param
documentation
}}
}
\value{
A list of integration weights minimizing the PRESS statistic,
the PRESS statistic itself, the post-integration estimates, and the
local-only estimates assuming no integration.
}
\description{
A joint fitting function to calculate integration weights for the full optimization and
2-parameter optimization methods as well as their post-integration weighted parameter
estimates.
}
