% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_fitting_survival.R
\name{joint_fitting_survival}
\alias{joint_fitting_survival}
\title{A joint fitting function to calculate integration weights for the full optimization and
2-parameter optimization methods as well as their post-integration weighted parameter
estimates.}
\usage{
joint_fitting_survival(
  datasets,
  fit.weighted = fit_weighted_survival,
  loocv.weighted = loocv_weighted_survival,
  test.coefs = test_coefs_survival,
  method = "testingpval",
  trace = F
)
}
\arguments{
\item{datasets}{A list of K datasets (each dataset having design matrix x, response y, and censoring vector censor).
It is assumed the first indexed dataset is the local dataset.}

\item{fit.weighted}{The fit_weighted_survival function}

\item{loocv.weighted}{The loocv_weighted_survival function}

\item{test.coefs}{The test_coefs_survival function}

\item{method}{One of:
\itemize{
\item testingpval - sets the integration weights equal to the likelihood ratio test (LRT) p-values
\item testingbinary - if the LRT is significant (p < 0.05), assign integration weight 0; 1 otherwise
\item opt - finds the length K-1 vector of integration weights maximizing the cross-validated partial likelihood
\item testingopt - using the better of testingpval or testingbinary as a starting point
find the 2-vector of p-value cutpoints that maximize the CVPL, turning a K-1 dimensional problem
into a 2 dimensional problem. For further details, refer to the loocv_weighted_survival_neg_2param
documentation
\item lasso - use glmnet to find the lambda minimizing the leave-one-out
cross validation error in a lasso-regularized Cox model, using a partial-likelihood loss
\item ridge - use glmnet to find the lambda minimizing the leave-one-out
cross validation error in a ridge-regularized Cox model, using a partial-likelihood loss
}}
}
\value{
For all non lasso/ridge methods, a list of integration weights minimizing the CVPL,
the CVPL, the post-integration estimates, and the local-only estimates assuming no integration.
For lasso and ridge, all of the above, as well as the penalty value lambda minimizing the
leave-one-out error as well as the LOO error value. Note that these methods assume no integration
and thus their vector of weights consists of all zeros.
}
\description{
A joint fitting function to calculate integration weights for the full optimization and
2-parameter optimization methods as well as their post-integration weighted parameter
estimates.
}
