% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_guo.R
\name{likelihood.guo}
\alias{likelihood.guo}
\title{The method of Guo et al., likelihood data fusion, implemented in R for comparison to our method}
\usage{
likelihood.guo(datasets, trace = F)
}
\arguments{
\item{datasets}{The list of K datasets (each dataset having design matrix x, response y)
It is assumed the first indexed dataset is the local dataset.}
}
\value{
If trace is off, returns parameters estimated by their method.
If trace is on, returns the above, as well as their integraiton weights and
individual coefficients for each dataset
}
\description{
The method of Guo et al., likelihood data fusion, implemented in R for comparison to our method
}
