% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_weighted_2param.R
\name{loocv_weighted_2param}
\alias{loocv_weighted_2param}
\title{Function to calculate the linear leave-one-out statistic, the PRESS statistic
This function is used for 2 parameter optmization as described in the paper}
\usage{
loocv_weighted_2param(alphabeta, datasets, precomputed.data = NULL)
}
\arguments{
\item{alphabeta}{The 2-vector of p-value cutpoints; any LRT p-value lower than alphabeta\link{1}
is given an integration weight of 0, any p-value higher than alphabeta\link{2} is given weight 1}

\item{datasets}{A list of datasets (each dataset having design matrix X, response y).
It is assumed the first indexed dataset is the local dataset.}

\item{precomputed.data}{Output from the precompute_linear function if already computed}
}
\value{
The linear leave-one-out cross-validation statistic for integrating the
external datasets with the given weights (the PRESS statistic). A lower value indicates a better model fit
}
\description{
Function to calculate the linear leave-one-out statistic, the PRESS statistic
This function is used for 2 parameter optmization as described in the paper
}
