% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_weighted_linear.R
\name{loocv_weighted_linear}
\alias{loocv_weighted_linear}
\title{Function to calculate the linear leave-one-out statistic, the PRESS statistic
This function is used for full optimization as described in the paper}
\usage{
loocv_weighted_linear(weights, datasets, precomputed.data = NULL)
}
\arguments{
\item{weights}{The vector of integration weights, should be equal to the # of external datasets}

\item{datasets}{A list of datasets (each dataset having design matrix X, response y).
It is assumed the first indexed dataset is the local dataset.}

\item{precomputed.data}{Output from the precompute_linear function if already computed}
}
\value{
The linear leave-one-out cross-validation statistic for integrating the
external datasets with the given weights (the PRESS statistic). A lower value indicates a better model fit
}
\description{
Function to calculate the linear leave-one-out statistic, the PRESS statistic
This function is used for full optimization as described in the paper
}
