% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_weighted_survival.R
\name{loocv_weighted_survival}
\alias{loocv_weighted_survival}
\title{Function to calculate the Cox leave-one-out statistic, the cross-validated partial likelihood (CVPL)}
\usage{
loocv_weighted_survival(weights, datasets, betafitted)
}
\arguments{
\item{weights}{The vector of integration weights, should be equal to the # of external datasets}

\item{datasets}{A list of datasets (each dataset having design matrix x, response y, and censoring vector censor).
It is assumed the first indexed dataset is the local dataset.}

\item{betafitted}{The post-integration weighted parameter estimates, found using fit_weighted_survival()}
}
\value{
The Cox leave-one-out cross-validation statistic for integrating the
external datasets with the given weights (the CVPL). A higher value indicates a better model fit
}
\description{
Function to calculate the Cox leave-one-out statistic, the cross-validated partial likelihood (CVPL)
}
