% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_weighted_survival_neg.R
\name{loocv_weighted_survival_neg}
\alias{loocv_weighted_survival_neg}
\title{Function to calculate the negative Cox leave-one-out statistic, the cross-validated partial likelihood (CVPL)
This function is passed to the base R function optim (which performs minimization) for the
"full optimization" described in the paper, hence it outputs the negative CVPL
(as the CVPL should be maximized).}
\usage{
loocv_weighted_survival_neg(weights, datasets)
}
\arguments{
\item{weights}{The vector of integration weights, should be equal to the # of external datasets}

\item{datasets}{A list of datasets (each dataset having design matrix x, response y, and censoring vector censor).
It is assumed the first indexed dataset is the local dataset.}
}
\value{
The negative Cox leave-one-out cross-validation statistic for integrating the
external datasets with the given weights (the CVPL). To be passed to optim().
}
\description{
Function to calculate the negative Cox leave-one-out statistic, the cross-validated partial likelihood (CVPL)
This function is passed to the base R function optim (which performs minimization) for the
"full optimization" described in the paper, hence it outputs the negative CVPL
(as the CVPL should be maximized).
}
