% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_weighted_survival_neg_2param.R
\name{loocv_weighted_survival_neg_2param}
\alias{loocv_weighted_survival_neg_2param}
\title{Function to calculate the negative Cox leave-one-out statistic, the cross-validated partial likelihood (CVPL)
This function is passed to the base R function optim (which performs minimization) for the
"2-parameter optimization" described in the paper, hence it outputs the negative CVPL
(as the CVPL should be maximized).}
\usage{
loocv_weighted_survival_neg_2param(alphabeta, datasets)
}
\arguments{
\item{alphabeta}{The 2-vector of p-value cutpoints; any LRT p-value lower than alphabeta\link{1}
is given an integration weight of 0, any p-value higher than alphabeta\link{2} is given weight 1}

\item{datasets}{A list of datasets (each dataset having design matrix x, response y, and censoring vector censor).
It is assumed the first indexed dataset is the local dataset.}
}
\value{
The negative Cox leave-one-out cross-validation statistic for integrating the
external datasets with the given weights (the CVPL). To be passed to optim().
}
\description{
Function to calculate the negative Cox leave-one-out statistic, the cross-validated partial likelihood (CVPL)
This function is passed to the base R function optim (which performs minimization) for the
"2-parameter optimization" described in the paper, hence it outputs the negative CVPL
(as the CVPL should be maximized).
}
