% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_linear.R
\name{precompute_linear}
\alias{precompute_linear}
\title{Function to precompute linear regression terms for integration speedup}
\usage{
precompute_linear(datasets, trace = F)
}
\arguments{
\item{datasets}{A list of K datasets (each dataset having design matrix X and response y).
It is assumed the first indexed dataset is the local dataset.}
}
\value{
Precomputed regression terms (X'X and X'y)
}
\description{
Function to precompute linear regression terms for integration speedup
}
