% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_survival.R
\name{run_simulation_survival}
\alias{run_simulation_survival}
\title{Simulations to assess the relative performance of the integration methods in reducing
estimation and prediction error}
\usage{
run_simulation_survival(
  K = 3,
  p = 5,
  n = c(100, rep(200, K - 1)),
  K0 = 1,
  sigma.beta = 1,
  mu.x = 0,
  sigma.x = 1,
  sigma.y = 0,
  seeds = c(1:6),
  numfrac = 0,
  trainingsize = 50,
  methods = c("opt", "testingopt", "lasso", "ridge"),
  correlated = F,
  equalcor = F
)
}
\arguments{
\item{K}{The total number of datasets (including the local dataset, which is assumed to be the first dataset)}

\item{p}{The number of parameters; for now hardcoded to 5 due to requirements of pec() function}

\item{n}{The sample size of the datasets, should be of length K}

\item{K0}{The number of datasets to integrate wrt the local dataset (aka have beta = beta0)}

\item{sigma.beta}{Additional noise to add to non-integrated or fractionally integrated datasets}

\item{mu.x}{Mean of the design matrix}

\item{sigma.x}{Standard deviation of the design matrix}

\item{sigma.y}{Noise to add to linear predictor in calculating exponentially distributed survival times}

\item{seeds}{Seeds to use to replicate simulations}

\item{numfrac}{Of the K0 to-be-integrated datasets, how many should be fractionally integrated
(i.e. beta has some noise)}

\item{trainingsize}{The size of the local dataset (sampled from the first dataset) to be
used as training data}

\item{methods}{Methods to run simulations for; refer to joint_fitting_survival documentation for further
information}

\item{correlated}{Designates whether the design matrix columns should be correlated (common in genomic data).}

\item{equalcor}{Only applies if correlated = TRUE. Designates if the correlations the same in all datasets.}
}
\value{
A matrix of improvements in estimation and prediction error, each column representing a specified
method. Values are presented as the log of the local-only error (no integration) divided by the post-
integration error; that is positive values represent improvements by performing integration.
}
\description{
Simulations to assess the relative performance of the integration methods in reducing
estimation and prediction error
}
