% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Function to simulate continuous, binary, or survival datasets}
\usage{
simulate_data(
  K = 3,
  p = 5,
  n = c(100, rep(200, K - 1)),
  beta0 = rnorm(p, 0, 1),
  c0 = 1,
  K0 = 1,
  sigma.beta = 1,
  mu.x = 0,
  sigma.x = 1,
  numfrac = 0,
  sigma.y = 1,
  correlated = F,
  equalcor = F,
  wt.frac = 0.8,
  heterosk = F,
  family = c("linear", "logistic", "cox")
)
}
\arguments{
\item{K}{The total number of datasets (including the local dataset, which is assumed to be the first dataset)}

\item{p}{The number of parameters}

\item{n}{The sample size of the datasets, should be of length K}

\item{beta0}{The "true" generating parameters}

\item{c0}{For survival data, governs times; greater c0 = shorter times}

\item{K0}{The number of datasets to integrate wrt the local dataset (aka have beta = beta0)}

\item{sigma.beta}{Additional noise to add to non-integrated or fractionally integrated datasets}

\item{mu.x}{Mean of the design matrix}

\item{sigma.x}{Standard deviation of the design matrix}

\item{numfrac}{Of the K0 to-be-integrated datasets, how many should be fractionally integrated (i.e. beta has some noise)}

\item{sigma.y}{Noise to add to linear predictor}

\item{correlated}{Designates whether the design matrix columns should be correlated (common in genomic data).}

\item{equalcor}{Only applies if correlated = TRUE. Designates if the correlations the same in all datasets.}

\item{wt.frac}{The noise to add to parameters of fractionally weighted external datasets:
noise ~ (1 - wt.frac) * N(0, 1)}

\item{heterosk}{Heteroskedasticity: makes response error a function of the design matrix}

\item{family}{The model to generate data for: linear = continuous data, logistic = binary data, cox = survival data}
}
\value{
A list containing a length vector of weights corresponding to parameter noise (weights,
NOT the integration weights); datasets each with the design matrix (x), response (y),
and for family "cox", censoring vector (censor); and the vector of true parameters beta0 (par)
}
\description{
Function to simulate continuous, binary, or survival datasets
}
