% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_coefs_linear.R
\name{test_coefs_linear}
\alias{test_coefs_linear}
\title{Function to conduct a likelihood ratio test (LRT), comparing the local dataset to external datasets}
\usage{
test_coefs_linear(datasets, trace = F)
}
\arguments{
\item{datasets}{A list of datasets (each dataset having design matrix X, response y).
It is assumed the first indexed dataset is the local dataset.}
}
\value{
A list of p-values corresponding to testing the null hypothesis that one shared set of parameters
is sufficient to describe the local dataset and each external dataset
}
\description{
Function to conduct a likelihood ratio test (LRT), comparing the local dataset to external datasets
}
