% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linear.R
\name{simulate_linear}
\alias{simulate_linear}
\title{Function to simulate multiple datasets for continuous data
This function simulates continuous data for integration with linear models}
\usage{
simulate_linear(
  K = 3,
  p = 5,
  n = c(50, rep(200, K - 1)),
  beta0 = rnorm(p, 0, 1),
  K0 = 1,
  sigma.beta = 1,
  mu.x = 0,
  sigma.x = 1,
  numfrac = 0,
  sigma.y = 0.5,
  correlated = F,
  equalcor = F,
  trace = F
)
}
\arguments{
\item{K}{The total number of datasets (including the local dataset, which is assumed to be the first dataset)}

\item{p}{The number of parameters}

\item{n}{The sample size of the datasets, should be of length K}

\item{beta0}{The "true" generating parameters}

\item{K0}{The number of datasets to integrate wrt the local dataset (aka have beta = beta0)}

\item{sigma.beta}{Additional noise to add to non-integrated or fractionally integrated datasets}

\item{mu.x}{Mean of the design matrix}

\item{sigma.x}{Standard deviation of the design matrix}

\item{numfrac}{Of the K0 to-be-integrated datasets, how many should be fractionally integrated (i.e. beta has some noise)}

\item{sigma.y}{Noise to add to linear predictor in calculating exponentially distributed survival times}

\item{correlated}{Designates whether the design matrix columns should be correlated (common in genomic data).}

\item{equalcor}{Only applies if correlated = TRUE. Designates if the correlations the same in all datasets.}
}
\value{
A list containing a length vector of weights corresponding to parameter noise (weights,
NOT the integration weights); datasets each with the design matrix (x), survival times (y),
and censoring vector (censor); and the vector of true parameters beta0 (par)
}
\description{
Function to simulate multiple datasets for continuous data
This function simulates continuous data for integration with linear models
}
